requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.29
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
mock==4.0.2
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
